refactor: move envvar out of internal tree so it can be imported externally
gomvpkg -from github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app/envvar -to github.com/sourcegraph/sourcegraph/cmd/frontend/envvar
