+++ cmd/frontend/internal/app/router/router.go
@@ -6,7 +6,7 @@ package router
 
 import (
 	"github.com/gorilla/mux"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app/envvar"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
 	"github.com/sourcegraph/sourcegraph/pkg/routevar"
 )
